pyproject.toml
[tool.poetry]
name = "try-object-detection"
version = "0.1.0"
description = "Example of object detection with MobileNet SSD v3"
readme = "README.md"
authors = ["Stepan Anokhin <stepan.anokhin@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
opencv-python = "4.5.4.60"
numpy = "^1.23.4"
tensorflow = "^2.10.0"
luigi = "^3.1.1"
matplotlib = "^3.6.1"
PyQt5 = "^5.15.7"
python-telegram-bot = "20.0a0"
httpx = "0.22.0"
orjson = "^3.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
prepare = 'try_object_detection.workflow.prepare_data:run'
detect = 'try_object_detection.workflow.detect_objects:run_file'
capture = 'try_object_detection.workflow.detect_objects:run_capture_cli'
bot = 'try_object_detection.telegram_bot.main:main'